OpenAI_Chatbot_Integration/requirements.txt
# Flask Web Framework
Flask==3.0.0

# OpenAI API Client
openai>=1.40.0

# HTTP Client (required by openai, pinned for compatibility)
httpx>=0.27.0

# Environment Variables Management
python-dotenv==1.0.0

# SQLAlchemy ORM & Flask integration
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0

# MongoDB Database (legacy, kept for backward compatibility)
pymongo==4.6.0
certifi>=2023.7.22

# Additional Dependencies
Werkzeug==3.0.1

# Production server for Cloud Run
gunicorn==21.2.0

# PostgreSQL adapter for Cloud SQL
psycopg2-binary==2.9.9